# docs/openapi/velontri_openapi.yaml
openapi: 3.1.0
info:
  title: Velontri API
  version: 1.0.0
  description: |
    Complete API for all 14 Velontri microservices.
    All responses use a standard envelope: { success, message, data, meta }
    All errors use: { success: false, error: { code, message, field }, request_id }
  contact:
    name: Velontri Engineering
    email: dev@velontri.com
servers:
  - url: https://api.velontri.com/api/v1
    description: Production
  - url: https://api.staging.velontri.com/api/v1
    description: Staging
  - url: http://localhost:8001/api/v1
    description: Auth service (local)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT. Access token TTL is 15 minutes.
  schemas:
    SuccessResponse:
      type: object
      required: [success, message]
      properties:
        success:  { type: boolean, example: true }
        message:  { type: string,  example: "Operation successful" }
        data:     {}
        meta:     { $ref: '#/components/schemas/PaginationMeta' }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:    { type: boolean, example: false }
        error:      { $ref: '#/components/schemas/ErrorDetail' }
        request_id: { type: string, format: uuid, example: "550e8400-e29b-41d4-a716-446655440000" }
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:    { type: string, example: "INVALID_CREDENTIALS" }
        message: { type: string, example: "Invalid email/phone or password." }
        field:   { type: string, nullable: true, example: "email" }
    PaginationMeta:
      type: object
      properties:
        page:        { type: integer, example: 1 }
        page_size:   { type: integer, example: 20 }
        total:       { type: integer, example: 150 }
        total_pages: { type: integer, example: 8 }
        has_next:    { type: boolean, example: true }
        has_prev:    { type: boolean, example: false }
...
    RegisterRequest:
      type: object
      required: [email, phone, password, full_name, country_code]
      properties:
        email:        { type: string, format: email,  example: "ada@example.com" }
        phone:        { type: string, pattern: '^\+[1-9]\d{6,14}$', example: "+2348012345678" }
        password:     { type: string, minLength: 8,   example: "SecurePass1!" }
        full_name:    { type: string, example: "Ada Okafor" }
        country_code: { type: string, minLength: 2, maxLength: 2, example: "NG" }
    LoginRequest:
      type: object
      required: [identifier, password, device_fingerprint]
      properties:
        identifier:          { type: string, example: "ada@example.com" }
        password:            { type: string, example: "SecurePass1!" }
        device_fingerprint:  { type: string, example: "web-chrome-abc123" }
        user_agent:          { type: string, example: "Mozilla/5.0" }
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
              example:
                success: true
                message: "Registration successful. Please verify your phone number."
                data: { user_id: "550e8400-e29b-41d4-a716-446655440000" }
        '422':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/verify-phone:
    post:
      tags: [Authentication]
      summary: Verify phone with OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id: { type: string, format: uuid, example: "550e8400-e29b-41d4-a716-446655440000" }
                otp:     { type: string, example: "123456" }
      responses:
        '200':
          description: Phone verified
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login with email/phone and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
              example:
                success: true
                message: "Login successful."
                data:
                  tokens:
                    access_token: "eyJhbGciOiJSUzI1NiJ9..."
                    refresh_token: "v2.local.AbCd..."
                    token_type: "bearer"
                    expires_in: 900
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/login/oauth:
    post:
      tags: [Authentication]
      summary: Login via Google or Apple OAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, id_token, device_fingerprint]
              properties:
                provider:           { type: string, enum: [google, apple] }
                id_token:           { type: string, example: "ya29.a0A..." }
                device_fingerprint: { type: string, example: "mobile-flutter-xyz" }
      responses:
        '200': { description: OAuth login successful }
  /auth/token/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: { type: string, example: "v2.local.AbCd..." }
      responses:
        '200':
          description: New access token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke refresh token
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: { type: string }
      responses:
        '200': { description: Logged out }
  /auth/introspect:
    get:
      tags: [Authentication]
      summary: Validate JWT and return claims
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Token valid
          content:
            application/json:
              example:
                data:
                  user_id: "550e8400-e29b-41d4-a716-446655440000"
                  roles: ["buyer", "seller"]
                  subscription_tier: "professional"
                  expires_at: 1720000900
  /auth/2fa/enable:
    post:
      tags: [Authentication]
      summary: Enable two-factor authentication
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                method: { type: string, enum: [totp, sms] }
      responses:
        '200': { description: 2FA enabled }
  /auth/2fa/verify:
    post:
      tags: [Authentication]
      summary: Complete 2FA challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                two_fa_session_id: { type: string }
                otp:               { type: string, example: "123456" }
      responses:
        '200': { description: 2FA verified, tokens returned }
  /auth/password/reset-request:
    post:
      tags: [Authentication]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email, example: "ada@example.com" }
      responses:
        '200': { description: Reset email sent (always returns 200 to prevent enumeration) }
  /auth/password/reset:
    post:
      tags: [Authentication]
      summary: Reset password using token from email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:        { type: string }
                new_password: { type: string, minLength: 8 }
      responses:
        '200': { description: Password updated }
  /auth/devices:
    get:
      tags: [Authentication]
      summary: List registered devices
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Device list }
  /auth/devices/{device_id}:
    delete:
      tags: [Authentication]
      summary: Revoke a device
      security: [{ BearerAuth: [] }]
      parameters:
        - name: device_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Device revoked }

  # ── MARKETPLACE ──────────────────────────────────────────────────────────────
...